+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.78
 ### is.codion.framework.model
 - AbstractEntityEditor, foreign key based detail editor no longer validates non-present detail entities.
+- AttributeOperands added, AbstractEntityTableModel.EntityFilters now initialize boolean conditions correctly.
 
 ## 0.18.77
 ### is.codion.common.reactive